requirements.txt
duckdb
dbt-duckdb
python-dotenv
requests
pandas